Initialize argon2i::salt to 0
The argon2i test was unstable because it only partially initialized
the non-random salt; I had been assuming that std::array's default
constructor initialized the contents to zero, but it doesn't.

For salt it seems a good idea to zero-fill since other uses cases
might only partially fill it, as the test does.

Fixes #2 (I think)