Moved Monocypher `crypto_` symbols out of global namespace
Build Monocypher itself as C++, wrapping it in a namespace.
This moves its `crypto_`-prefixed symbols out of the global
namespace where they can collide with libSodium or OpenSSL.

Note: This required a small change to Monocypher's headers.
I've made the change in a fork, and updated the .gitmodules
file to point to the fork.